Added proper support for the double rotor frame order models to the system test scripts.
This is for the CaM synthetic data.  The base script can now handle the current parameterisation of
the double rotor model with a single pivot, an eigenframe, and the second pivot defined by a
displacement along the z-axis.  The scripts for the double_rotor and double_rotor_large_angle data
sets have been changed to use this parameterisation as well.