Created a master profiling script for comparing the speed between different dispersion models.
This is similar to the disp_profile_all.py script except it only operates on a single relax version.
The output is then simply the timings, with statistics, of the calculation time for 100 function
calls for 100 spins (either 10,000 function calls for single spins or 100 function calls for the
cluster of 100 spins).

The output of the script for the current disp_spin_speed branch code has also been added.