Improvements to the internal structural object _get_chemical_name() method.
This now uses a translation table to convert the hetID or residue name into a description, for
example as used in the PDB HETNAM records to give a human readable description of the residue inside
the PDB file itself.  The new rotor RTX and RTB residue names have been added to the table as well.