Renaming of the residues of the cone geometric representation.
The cone apex or centre is now the CNC residue, the cone axis is now CNX and the cone edge is now
CNE.  These used to be APX, AXE, and EDG respectively.  The aim is to make these names 100% specific
to the cone object so that they can be more easily selected for manipulating the representation and
so that they are more easily identifiable.  The internal structural object _get_chemical_name()
function now returns a description for each of these.  Note that the main cone object is still named
CON.