The motional pivots for the frame order models are now labelled in the geometric representation.
The pivot points are now added as a new molecule called 'pivots' in the frame_order.pdb_model user
function.  The atoms all belong to the PIV residue.  The pymol.frame_order user function now selects
this residue, hides its atoms, and then shows the atom name 'Piv' as the label.  For the double
rotor model, the atom names 'Piv1' and 'Piv2' are used to differentiate the pivots.