Expansion of the REMARK section of the PDB file created for the internal structural object.
This is visible when using the structure.write_pdb user function, as well as the many other user
functions which create PDB files.  The relax version as well as the file creation date are now
recorded in the PDB file.  This extra information should be very useful.  Empty lines in the REMARK
section improve the formatting.