Started to redesign the frame_order.pdb_model user function.
Instead of having the positive and negative representations in different PDB models, and the Monte
Carlo simulations in different molecules, these will now all be shifted into separate files.

For this to be possible, the file root rather than file names must now be supplied to the
frame_order.pdb_model user function.  To allow for different file compression, the compress_type
argument is now used.  The backend code correctly handles the file root change, but the multiple
files are not created yet.