Redesign of the average domain position part of the frame_order.pdb_model user function.
The Monte Carlo simulations are now represented.  If the file root is set to the default of
'ave_pos', then these will be placed in the file 'ave_pos.pdb', or a compressed version.  Each
simulation is in a different model, matching the geometric representation '*_sim.pdb' files.  The
original structure is copied for each model, and then rotated to the MC simulation average position.