The PyMOL disable command is now used by the pymol.frame_order user function.
This is to first disable all PyMOL objects prior to loading anything, to hide the original
structures and any previous frame order representations, and then to hide all of the Monte Carlo
simulation representations.  This is to simplify the picture initially presented to the user while
still allowing all elements to be easily found.