Added titles to the frame order geometric representation PDB files from frame_order.pdb_model.
These are in the form of special Ti atoms placed 40 Angstrom away from the pivot along the z-axis of
the system, or shifted 3 more Angstrom for the Monte Carlo simulations.  These are used to label the
alternative representations or the Monte Carlo simulation representations.  The residue type is set
to TLE and this has been registered in the internal structural object.

The pymol.frame_order user function now calls the represent_titles() function to select these atoms,
hide them, and then add a long descriptive title.  The atom name is used to distinguish between
different titles.