Changed the alternative representation names for the frame order geometric objects.
The aim is to put both representations on a more equal footing, as they are identical solutions.
Hence the inverted representation might be the correct representation of the domain motions.  So
instead of calling these 'positive' and 'negative', the 'A' and 'B' notation will be used.

This affects the names of the files produced by the frame_order.pdb_model user function as well as
the internal titles.  Instead of ending the files with "*_pos.*" and ""*_neg.*", these have been
changed to "*_A.*" and ""*_B.*".  The atoms used for the titles have also been renamed, and the
pymol.frame_order user function now labels the titles using the 'A' and 'B' notation.