Redesign of the geometric representation of the cone structural objects to allow for models.
The old representation was not compatible with the PDB model concept whereby each model must have
the same number of atoms.  To handle this situation, the cone objects have been simplified.
Specifically the cone cap.

The old behaviour was to remove all points outside of the cone when creating the cone cap, and then
to stitch the cap to the cone edge in a subsequent step.  Now the behaviour is that all points
outside of the distribution are shifted to the cone edge.  This avoids the need to stitch the cap to
the edge.  This behaviour means that all cones with the same inc value will have the same number of
atoms.

The cones for the pseudo-ellipses are not as nice as the latitudinal lines are not strait at the
cone edge, but at least creating multiple models with different cone sizes is now possible.