Changed the CaM frame order test data superimposition values.
Because the domains are now defined via the molecule name rather than the residue numbers, the
centroid of rotation set to the CoM has been shifted as now the Ca2+ ions are included in the CoM
calculation.  Therefore the superimpose.py script has been updated to not delete the Ca atoms (this
reverts r23354).

All of the frame order optimisation scripts have been updated with the new rotation Euler angles and
translation vector.  To match this, the system test base script for the CaM frame order test data
has also had its rotations and translations updates, and the domain user function call changed to
use molecule names.