Fix for the relaxation dispersion analysis Monte Carlo simulation printouts on clusters.
The multi-processor code was calling the print() function from the Slave_command.run() method,
however this runs on the slave processor.  This has been shifted to the Results_command.run() method
which runs on the master once the results have been returned via the Results_command.  Now the
printout of the simulation number and cluster ID will be visible when running via OpenMPI on a
cluster.