Parallelised the frame order grid search to run on clusters or multi-core systems via OpenMPI.
This involved the creation of the Frame_order_grid_command class which is the multi-processor
Slave_command for performing the grid search.  This was created by duplicating the
Frame_order_minimise_command class and then differentiating both classes.

For the subdivision of the grid search, the new minfx grid.grid_split_array() function is used in
the frame order grid() API method.  The grid() method no longer calls the minimise() method but
instead obtains the processor box itself and adds the subdivided grid slaves to the processor.  The
relax grid_search user function takes care of the rest.