Fixes for the parallelised grid search for the frame order analysis.
A chi-squared value check was added to the Frame_order_result_command.run() method to check if the
value is lower than the current when the result is returned to the master.  Without this check, each
grid subdivision result will be stored as they are returned rather than storing the results from the
global minimum of the entire grid search.