Improvements for the parallelised grid search for the frame order analysis.
As each grid point can take wildly different numbers of CPU cycles to calculate the chi-squared
value for, the result of subdividing the grid search was that some subdivisions where incredibly
quick while others required much larger amounts of time.  To avoid this bad slave management, the
grid points are now randomised.  This means that the subdivisions will require about the same amount
of time to optimise.