Moved the setup of the target function data structures in the frame order analysis.
This is for the grid_search and minimise user functions.  The target function data setup function
has been renamed to target_fn_data_setup().  This is now called before the Frame_order_grid_command
and Frame_order_minimise_command multi-processor objects are initialised, and all of the data is now
passed into these functions.  Although the code is uglier, this has the benefit that the
target_fn_data_setup() function will only be called once.  This data setup requires a lot of time,
so for a large cluster, this can be a large time saving for the grid search.