Redesign of the optimisation user functions calculate, grid_search, and minimise.
In preparation for expanding the number of optimisation user functions, these three current user
functions have been shifted into the new minimise user function class.  The calculate user function
is now accessed as minimise.calculate.  The grid search as minimise.grid_search.  And minimisation
is via the minimise.execute user function.  The icon used for the new user function class is the
Rosenbrock function or the banana optimisation problem.

As this is such a radical change, a huge number of changes in the relax source code, the sample
scripts, the user manual, and the test suite were required.