Created the parameter object infrastructure for adding the grid search lower and upper bounds.
The _add() method now accepts the grid_lower and grid_upper keyword arguments, which can be either
values or functions.  These are then stored in the _grid_lower and _grid_upper class dictionaries.
The public methods grid_lower() and grid_upper() have been added to return the value corresponding
to the given parameter.