Updated all of the specific API grid_search() methods to accept the scaling_matrix argument.
The list of per-model scaling matrices is now created independently of the analysis type by the
pipe_control.minimise methods for the minimise.calculate, minimise.grid_search and minimise.execute
user functions and sent into the specific analysis backend.

The argument is also passed into the minimise() API method from the grid_search() method when that
is used.