Major change to the grid search user function.
The minimise.grid_search user function now accepts the skip_preset flag.  When True, the grid search
will skip any parameters with a preset value.  This allows the user to set parameters via the
value.set user function and then have these parameters skipped in the grid search.

The new skip_preset argument is passed into the pipe_control.minimise.grid_setup() function in the
backend.  This then sets both the grid lower and upper bounds to the preset parameter value and sets
the number of increments to 1 for that parameter so that it is essentially skipped in the grid
search.