Converted the relaxation dispersion analysis to the new grid bounds and scaling factor design.
The parameter object now registers the grid bounds and scaling factors for all of the dispersion
parameters.  This includes the functions dw_lower(), dwH_lower(), pA_lower() and i0_upper() in the
specific_analyses.relax_disp.parameter_object module for calculating some of these values.

The specific_analyses.relax_disp.parameters.assemble_scaling_matrix() function has been deleted as
this is now provided by the upstream code in pipe_control.minimise.  And the API methods
grid_search() and minimise() has been modified to accept the list of scaling matrices.

As the grid bounds and increments are now handled by the upstream pipe_control.minimise.grid_setup()
function, the specific_analyses.relax_disp.optimisation.grid_search_setup() function was redundant
and was deleted.

The specific_analyses.relax_disp.parameters.get_param_names() function was also modified with the
full argument added, to allow either the base parameter names or an augmented form with the
dictionary key for presenting to the user to be returned.

Importantly to allow the changes to be operational, the model_loop() API method was redesigned so
that, for the R2eff base model, the individual spins rather than spin clusters will be looped over.
This allows the specific_analyses.relax_disp.optimisation.minimise_r2eff() function to continue to
operate correctly.