The grid search upper and lower bound functions now must accept the incs argument.
For a few analyses, the number of grid search increments are used to remove the end point of the
grid, to remove duplicate points due to the circular nature of angles.  Therefore the parameter
object grid_lower() and grid_upper() methods now send the grid increment number for each parameter
into all grid bound determining functions.  The relaxation dispersion and curve-fitting analyses
have been updated for the change.