Modified all calls to the parameter object _add() base class method.
These are now all spread across multiple lines, with each argument on a separate line.  This is for
easier maintenance of the specific analysis parameters, as the code is now much cleaner and argument
changes will only have diffs for that argument.  It is also visually easier to see all the settings
for each parameter.