Completed the Relax_fit.test_zooming_grid_search system test.
Now only a single spin is optimised.  The zooming levels increase in integer increments from 0 to 50
so that the final zoomed grid is insanely small (as the curve-fitting C modules are incredibly fast,
this test is nevertheless relatively quick).  The final zooming grid search parameter values are
checked to see if they are the same as those optimised in the Relax_fit.test_curve_fitting_height
system test to demonstrate the success of the algorithm.