Modified the grid search upper bounds functions for the relaxation curve-fitting.
This is for both the exponential relaxation curve-fitting analysis and the same fitting in the
dispersion analysis.  The intensity values are doubled and then rounded to the next order.  This
ensures that I0 and Iinf will be within the grid bounds.  Hence the zooming grid search can be used
for these curves.