Added a 128x128 pixel version of the zoom-in Oxygen icon.
This icon size is not available in the repository located at
svn://anonsvn.kde.org/home/kde/trunk/kdesupport/oxygen-icons.  Therefore the
scalable/actions/small/48x48/zoom-in.svgz file was copied and then exported into a 128x128 PNG, and
finally converted to a Gzipped EPS file for the relax manual.