Redesign and expansion of the nested model parameter copying in the frame order auto-analysis.
The nested parameter protocol used to allow the analysis to complete in under 1,000,000 years was no
longer functional due to the switching to the axis alpha parameter to decrease parameter number and
redundancy.  The copying of the average domain position for the free rotor models was also incorrect
as the dropping of the alpha Euler angle cause the translation parameters and beta and gamma angles
to change drastically.

The new protocol has been split into four methods for the average domain position, the pivot point,
the motional eigenframe and the parameters of ordering.  These use the fact that the free rotor and
torsionless models are the two extrema of the models where the torsion angle is restricted.  The
pivot copying is a new addition.