Implementation of the Optimisation_settings object in the frame order auto-analysis.
This object holds all of the grid search, zooming grid search, and minimisation settings.  It
provides the add_grid() and add_min() methods to allow the user to add successive iterations of
optimisation and settings to the object.  The loop_grid() and loop_min() methods are used to loop
over each iteration of each method.  And the get_grid_inc(), get_grid_num_int_pts(),
get_grid_zoom_level(), get_min_algor(), get_min_func_tol() and get_min_num_int_pts() methods are
used to access the user defined settings.

The auto-analysis has been redesigned around this new concept.  All of the optimisation arguments
have been replaced.  Instead there are the opt_rigid, opt_subset, opt_full, and opt_mc arguments
which are expected to be instances of the Optimisation_settings object.  The optimisation in the
auto-analysis is now more advanced in that more user optimisation settings are now available and
active.