The frame order auto-analysis Optimisation_settings object now handles the maximum iterations.
The new max_iter argument has been added to the add_min() method, and the new get_min_max_iter()
method added to fetch the value.  This is used in the auto-analysis to set the maximum number of
optimisation iterations in the minimise.execute user function calls.  Limiting this will be of
greatest benefit for the test suite.