Removed the RDC data checks from the frame order optimisation.
This is in the minimise_setup_rdcs() and store_bc_data() functions of the
specific_analyses.frame_order.optimisation module, called before and after all optimisation.  The
reason was identified by profiling - this check was adding significant amounts of time to the setup
and results unpacking parts of the optimisation.  Specifically the interatomic_loop() function was
identified via profiling as the function requiring the most amount of cumulative time in the
Frame_order.test_auto_analysis system test (17 seconds out of a total of ~60 seconds).