Created function to create the helper index numpy array, to help figuring out the indices to store in the exchange data matrix.
This is for striding through data and store the data correct in the data matrix.

This is for a special situation where numpy version is < 1.8, and where the numpy.linalg.eig() function can only
be performed on square matrices, and not on higher dimensional data.

For this situation, it is necessary to stride through the data.