Implemented partial reading of results file.
Before reading a results file, it is determined if the file exists.

This makes is possible to read a directory with partial results from a previous analysis.

This can be handsome, if reading R2eff values in R1rho experiments, and the error estimation has been
prepared with a high number of Monte-Carlo simulations.

sr #3135(https://gna.org/support/?3135): Optimisation of the R1 relaxation rate for the off-resonance R1rho relaxation dispersion models.