Removed the auto-sorting of models, when performing auto analysis of Relaxation dispersion.
This was discussed in:
http://thread.gmane.org/gmane.science.nmr.relax.scm/22733
http://thread.gmane.org/gmane.science.nmr.relax.scm/22734
http://thread.gmane.org/gmane.science.nmr.relax.scm/22737

Through this discussion, it has appeared that the order of how models are sorted for analysis, and hence the
possibility for nesting, is a complicated case.

The order of analysis should be possible to manually put into the auto analysis.

This was not the scope of:
sr #3135(https://gna.org/support/?3135): Optimisation of the R1 relaxation rate for the off-resonance R1rho relaxation dispersion models.
to implement such a feature.

Such a feature could be implemented for the next version of relax.

It could be designed as function to "suggest" an order in the GUI.

But this functionality would have to wait.