Removed results files to allow the Relax_disp.test_r1rho_kjaergaard_missing_r1 system test to pass.
These are the test_suite/shared_data/dispersion/Kjaergaard_et_al_2013/check_graphs/mc_2000/ results
files for the 'No Rex' and 'DPL94' models, as well as the final run.

This commit is to allow the test to temporarily pass.  It can be reverted once a better solution is
discussed and decided upon.