Merger of the 'No Rex' and 'No Rex R1rho off res' models in the specific_analyses.relax_disp package.
In the 'variables' module, all *_NOREX_R1RHO variables have simply been deleted and the MODEL_LIST_*
structures updated.

For the 'data' module, the is_r1_optimised() function was modified to catch the 'No Rex' model and
to then use the cdp.exp_type_list structure to determine if the experiment type is EXP_TYPE_R1RHO.
This will be modified in the future by using a function for determining if the current experiment is
on or off-resonance.  The return_r1_data() and return_r1_err_data() functions have also been
modified to check if R1 values are fit rather than if the model is in MODEL_LIST_R1RHO_OFF_RES.

In the 'model' module, in addition to deleting all *_NOREX_R1RHO variables, the convert_no_rex()
function has also been deleted as it no longer serves a purpose.

In the 'checks' module, all 'No Rex R1rho off res' model references have been replaced with 'No
Rex'.