Fix for bug #22502 (https://gna.org/bugs/?22502).
This is the problem whereby the geometric prolate diffusion representation does not align with axis
in PDB, as reported by Martin Ballaschk (https://gna.org/users/mab).

This problem was not the main prolate tensor axis, but that the geometric object needed to be
rotated 90 degrees about the z-axis to bring the object and axis into the same frame.