Implementation of the target_functions.relax_fit.jacobian() function.
This follows from the discussions at http://thread.gmane.org/gmane.science.nmr.relax.devel/6807.

The function will calculate the Jacobian matrix for the exponential curve-fitting module.  The
Jacobian can be used to directly calculate the covariance matrix, for example as described at
https://www.gnu.org/software/gsl/manual/html_node/Computing-the-covariance-matrix-of-best-fit-parameters.html.
The Jacobian is calculated using the help of the new exponential_dI() and exponential_dR() functions
in the target_functions/exponential.c file.  These calculate the partial derivatives of the
exponential curve with respect to each model parameter separately.

The implementation still needs testing and debugging.