Major bugfix for the frame order auto-analysis.
The algorithm of using a PCS data subset of a few selected residues to find an initial parameter
estimate followed by using all PCS data was badly implemented.  The use of the PCS subset caused
most spin systems to be deselected, however they remained deselected once all data was being used.
So the result was that only the spin subset was ever being used in the analysis.