Implemented the specific_analyses.relax_fit.optimisation.d2func_wrapper() function.
This converts the numpy parameter array into a Python list, calls the
target_functions.relax_fit.d2func() function with this list, converts the Hessian output list of
lists into a numpy rank-2 array, and returns it.  This will allow Newton optimisation to be used for
the relaxation curve-fitting analysis.