The minimisation statistics are now being reset for all analysis types.
The minimise.calcululate, minimise.grid_search, and minimise.execute user functions now all reset
the minimisation statistics for either the model or the Monte Carlo simulations prior to performing
any optimisation.  This is required for both parallelised grid searches and repetitive optimisation
schemes to allow the result to overwrite an old result in all situations, as sometimes the original
chi-squared value is lower and the new result hence is rejected.

The equivalent functionality from the model-free minimise() API method has been removed.