By using minfx, and the reported Jacobian, it is now possible to get the exact same error estimation as scipy.optimize.leastsq.
The fatal error was to set the weighting matrix with diagonal elements as the error.
There weights are 1/errors**2.

There is though some un-answered questions left.

The Jacobian used, is the direct derivative of the function.

It is not the chi2 derivative Jacobian.

task #7822(https://gna.org/task/index.php?7822): Implement user function to estimate R2eff and associated errors for exponential curve fitting.