Removed the user function to estimate the R2eff values and errors with scipy.optimize.leastsq.
With the newly implemented Jacobian and Hessian of the exponential decay function, the front-end to
scipy.optimize.leastsq does not serve a purpose.

This is because minfx is now as fast as scipy.optimize.leastsq, and can estimate the errors from the
Jacobian to the exact same numbers as scipy.optimize.leastsq.

In addition to that, the covariance can be calculated by QR decomposition.
This adds additional feature for checking for a singular matrix.

The back-end wilÃl still be kept in place for the coming tim, but could be removed later.

task #7822(https://gna.org/task/index.php?7822): Implement user function to estimate R2eff and associated errors for exponential curve fitting.