Added systemtest, Relax_disp.test_estimate_r2eff_err_auto and extended functionality to the auto analyses protocol.
If "exp_mc_sim_num" is set to "-1" and sent to the auto analyses, the errors of R2eff will be estimated from the Co-Variance matrix.

These errors is HIGHLY likely to be wrong, but can be used in an initial test fase, to rapidly produce data for plotting data.

task #7822(https://gna.org/task/index.php?7822): Implement user function to estimate R2eff and associated errors for exponential curve fitting.