Added keyword "exp_mc_sim_num", to the auto analyses in the GUI.
This sets the number of Monte Carlo simulations for R2eff error estimation in exponential curve fitting.

When setting to -1, the errors are estimated from the Covariance Matrix.

These errors are highly likely to be wrong, but can be used in Rapid testing of data and plotting.

task #7822(https://gna.org/task/index.php?7822): Implement user function to estimate R2eff and associated errors for exponential curve fitting.