Created the lib.periodic_table.process_symbol() function.
This will take an atomic symbol and return a copy of it with an uppercase first letter and lowercase
second letter.  This is used by the Periodic_table methods atomic_mass() and atomic_weight() to
allow for non-standard symbol input, for example if the element name comes directly from the all
uppercase PDB file format without translation.