First attempt at properly implementing the target_functions.relax_fit.jacobian() function.
This is now the Jacobian of the chi-squared function.  A new jacobian_matrix data structure has been
created for holding the matrix data prior to converting it into a Python list of lists.  The
equation used was simply the chi-squared gradient whereby the sum over i has been dropped and the i
elements are stored in the second dimension of matrix.