Speed up of the target_functions.relax_fit C module.
The variances are now precalculated in the setup() function from the errors, so that the use of the
square() function is minimised.  The chi-squared equation, gradient, and Hessian functions now
accept the variance rather than standard deviation argument and hence the squaring of errors has
been removed.  This avoids a lot of duplicated maths operations.