An attempt at documenting the Monte Carlo simulation verses covariance matrix error estimates.
This is for the R2eff and I0 parameters of the exponential curves.  For the Monte Carlo errors,
10000 simulations were preformed.  This means that these errors can perform as a gold standard by
which to judge the covariance matrix technique.

Currently it can be seen that the relax_disp.r2eff_err_estimate user function with the chi2_jacobian
flag set to True performs extremely poorly.