Comparison of 10,000 Monte Carlo simulations to a different covariance matrix error estimate.
The covariance_matrix.py script has been duplicated and the chi2_jacobian argument of the
relax_disp.r2eff_err_estimate user function has been changed from True to False.  As can be seen in
the 2D Grace plots, this error estimate is incredibly different.

The R2eff errors are overestimated by a factor of 1.9555, which indicates that the Jacobian or
covariance matrix formula are not yet correct.