Tried to implement bootstrap method.
But that goes horrible wrong.

The numbers are now:
param: r2a_err, with err: 0.48131663, compared to MC: 0.49354342, with 600 boot 0.00000000
param: dw_err, with err: 117.99734537, compared to MC: 0.32813703, with 600 boot 0.00000000
param: k_AB_err, with err: 0.64606994, compared to MC: 0.65384783, with 600 boot 0.00000000

So, something is right, and something is really wrong.

It is clear, that the reported error estimation scales with the error.

So, either one should start with simulations from peak intensity and then forward.
Or, maybe the use of errors are wrong.

There are some rules for how errors can be added together.
If they are a product, then the fractional errors can be summed together.

But, I have to look into this.

task #7824(https://gna.org/task/index.php?7824): Model parameter ERROR estimation from Jacobian and Co-variance matrix of dispersion models.