Change to the parameter nesting in the frame order auto-analysis.
The pivot is now taken from the rotor model for all other models.  Taking the pivot point from the
isotropic cone model is not a good idea as there are situations where the pivot point optimisation
catastrophically fails, sending the point many tens or hundreds of Angstrom away from the molecule.