Another fix for the minfx version checking in the dep_check module.
The version_comparison() function has been created to perform a proper version number comparison by
stripping trailing zeros, converting the two version numbers to lists of int and comparing the lists
using the Python cmp() function.  This will return -1 when the version number is too low, 0 when the
versions are equal, and 1 when the version is higher than the minimum.