Added dataset which was produced with BFGS.
This shows, that gradient calculation is either wrong in relax, or in minfx.
This is only for minimisation.

It is also easily to see on the graphs.

param: r2a_err, with err: 0.70912495, compared to MC: 0.10395753, with 500 boot 0.66165472
param: dw_err, with err: 0.56449802, compared to MC: 0.03708529, with 500 boot 0.00703459
param: k_AB_err, with err: 0.77161279, compared to MC: 0.29434628, with 500 boot 0.79263287

task #7824(https://gna.org/task/index.php?7824): Model parameter ERROR estimation from Jacobian and Co-variance matrix of dispersion models.