Modified for systemtest with BFGS.
Data was created with simplex, minimies with simplex but MonteÃÃÃ-Carlo simulations was with BFGS.

The error estimation is still the same.
So this tells, that minimisation is wrong in relax with BFGS.

Either minfx suffer, or my setup of the chi2 derivative suffers.

With simplex
param: r2a_err, with err: 0.48131663, compared to 2000 MC: 0.49354342, with 200 boot 0.43066033
param: dw_err, with err: 0.30874093, compared to 2000 MC: 0.32813703, with 200 boot 0.30946606
param: k_AB_err, with err: 0.64606994, compared to 2000 MC: 0.65384783, with 200 boot 0.62079033

With BFGS
param: r2a_err, with err: 0.47885323, compared to 2000 MC: 0.12582452, with 200 boot 0.41553408
param: dw_err, with err: 0.30068501, compared to 2000 MC: 0.07477556, with 200 boot 0.00507689
param: k_AB_err, with err: 0.64694369, compared to 2000 MC: 0.40271214, with 200 boot 0.64937810

task #7824(https://gna.org/task/index.php?7824): Model parameter ERROR estimation from Jacobian and Co-variance matrix of dispersion models.