The prompt UI now uses the user_functions.uf_translation_table dictionary.
The modified runcode() function will now check if the command typed by the user is a function or
method call and then will raise a RelaxError if the command name is in the
user_functions.uf_translation_table dictionary, telling the user that the user function has been
renamed to the new name in the translation table.

This appears to have no effect in the script UI however.