To the back-end of display pipes, added functionality to sort the pipe names before printing.
Also added the return of the list of pipes, with its associated information about pipe type, and pipe_bundle.

This is to help with getting a better overview for multiple pipes in data store.

task #7826(https://gna.org/task/index.php?7826): Write an python class for the repeated analysis of dispersion data.