Added support for axis permutations in the frame order auto-analysis.
This is done by copying the data pipe of the already optimised pseudo-elliptic models, permuting the
axes, and performing another optimisation using all RDC and PCS data.  This allows the second
solution for these pseudo-elliptic models to be found.  The 2nd pipe is included in the model
selection step to allow the best solution for the model to be found.