Implemented the new frame_order.permute_axes backend.
The 3 starting conditions x<y<z, x<z<y, and z<x<y and the two permutations 'A' and 'B' (for each of
these which do not include the starting permutation) are now supported.  For these 6 combinations,
the axis and order parameter permutation and the z-axis inversion are selected and applied to the
current system.