Added support for the isotopic cone models to the frame_order.permute_axes user function.
This is a simpler setup, but it uses the same permutation algorithm as derived for the
pseudo-ellipse models.  Instead of setting the x and y cone angles separately, they are instead
averaged.  And as the cone axis is undefined in the xy plane, the axis has been randomly selected as
being the axis perpendicular to both the z-axis and the reference frame x-axis.