Created set of files showing the axis permutation problem for the isotopic cone frame order model.
This shows that there are two minima.  However one has a chi-squared value of ~1, and the other a
value of ~150.  Nevertheless, the optimisation could be trapped in the non-global minimum so the
frame_order.permute_axes user function should be used for the isotopic cones as well, just in case.