Shifted the contents of the specific_analysis.relax_fit.estimate_rx_err module into the API.
The estimate_rx_err() function is now the covariance_matrix() method of the specific API.  The code
for calculating the covariance matrix and errors are now in the function
pipe_control.error_analysis.covariance_matrix(), so this has been removed.  And the error setting is
performed by the set_errors() API method, so that code has been deleted as well.