Added support for the model argument for the frame_order.pdb_model user function.
This argument is used to specify which of the models in an ensemble will be used to represent the
average domain position Monte Carlo simulations, as each simulation is encoded as a model, as well
as for the distribution of structures simulating the motion of the system.  The argument is
therefore passed into the create_ave_pos() and create_distribution() functions of the
specific_analyses.frame_order.geometric module.

To handle all models being used in the non Monte Carlo simulation PDB file and only one in this
file, the internal structural object is copied twice.  The second copy for the MC sims has all but
the chosen model deleted out of it.