The frame_order.pdb_model user function can now properly handle a failed pivot optimisation.
This is when the pivot point optimises to a coordinate outside of the PDB limits.

Now all calls to specific_analyses.frame_order.data.generate_pivot() from the module
specific_analyses.frame_order.geometric set the pdb_limit flag to True.  This allows all
representation objects to be within the PDB limits.  The algorithm in generate_pivot() has been
extended to allow higher positive values, as the real PDB limits are [-999.999, 9999.999].  And a
RelaxWarning is called when the pivot is outside to tell the user about it.