The frame order count_sobol_points() function is now being called by all of minimise user functions.
This occurs at the end of the minimise.calculate, minimise.grid_search, and minimise.execute user
function backends to provide more feedback to the user as to the quality of the optimisation.  To
avoid initialising the target function twice, the count_sobol_points() function now accepts the
initialised target function as an optional argument.