Redesigned the frame_order.num_int_pts user function frontend for the oversampling idea.
The use of the quasi-random Sobol' sequence for numerical PCS integration will be modified to use
the concept of oversampling.  Instead of specifying the exact number of points in the Sobol'
sequence and then removing points outside of the current parameter values, the algorithm will
oversample as N * Ov * 10**M, where N is the maximum number of Sobol' points to be used for the
integration, Ov is the oversampling factor, and M is the number of dimensions or torsion-tilt angles
used in the system.  The aim is to try to use the maximum number of points N for all frame order
models and all ranges of dynamics.