Changed the creation of the Sobol' points in the frame order target function.
For increased accuracy of the numerical PCS integration, the first 1000 points of the Sobol'
sequence are now skipped to avoid any bias.  For speed, the axis order of the Sobol' torsion-tilt
angles has been swapped so that the numpy.swapaxes() function call is no longer required in the
lib.frame_order.*.pcs_numeric_int_*() functions.