Huge speed up for the generation of the Sobol' sequence data in the frame order target function.
The new Sobol_data class has been created and is instantiated in the module namespace as
target_function.frame_order.sobol_data.  This is used to store all of the Sobol' sequence associated
data, including the torsion-tilt angles and all corresponding rotation matrices.  When initialising
the target function, if the Sobol_data container holds the data for the same model and same total
number of Sobol' points, then the pre-existing data will be used rather than regenerating all the
data.  This can save a huge amount of time.