Created a special internal structural object selection object.
This will be used for massively speeding up the internal structural object.  The use of the
lib.selection module by the internal structural object is currently very slow as a huge number of
calls to re.search() are required.  The idea is to avoid this by using lib.selection once to
populate this new selection object, and then reusing this object to loop over molecules and atoms.