Converted all structural object methods to use the selection object rather that atom ID strings.
This should have a significant impact on the speed of certain operations within relax.  The most
obvious effect will be a huge speed up of the interatom.define user function.  There should be speed
ups with a number of other user functions relating to structural information.  All parts of relax
have been updated for the change.