Fixes for the missing cdp.quad_int flag.
If the cdp.quad_int flag is missing, this is now set to False before setting up the target function
class.  The previous behaviour was that the frame_order.quad_int user function must be called prior
to optimisation.  Now it is optional for turning this flag on and off.