Clean up for the frame order target function aliasing.
The Scipy quadratic integration and the quasi-random Sobol' integration target functions are now
aliased using the getattr() Python method to programmatically choose one or the other.  The rigid
model has been removed from the list as it is not a numeric model, and the func_double_rotor()
target function has been renamed to func_double_rotor_qrint() to make it consistent with the naming
of the other target functions.