Updated all of the chi2 value checks for the Frame_order.test_cam_quad_int_* system tests.
This is only for those tests which use PCS data - the RDC only test chi2 values are the same as in
the Frame_order.test_cam_qr_int_* system tests.  In all cases, the chi2 value is lower for the more
accurate SciPy quadratic integration as compared to the quasi-random Sobol' integration, as
expected.