Added the ability to specify a pre-run directory in the frame order auto-analysis.
This will be used for refinement purposes.  If the new pre_run_dir argument, modelled on the
relaxation dispersion auto-analysis, is supplied then results files will be loaded from this
directory and the base data pipe copying and PCS subset optimisation steps will be skipped.  The
model nesting algorithm is also deactivated.