Elimination of most of the Frame_order.fixme_test_* system tests and associated data.
These tests are from a very early stage of the development of the frame order theory back when the
base data was the full and reduced alignment tensors for the each domain calculated from the RDC
data.  They do not fit into the current analysis where the base data is the RDCs and PCSs for the
moving domain.  There is no point upgrading the tests as it will be far too much effort and it will
only duplicate the coverage of the Frame_order.test_cam_* system tests.