Fix for the frame order specific API calculate() method.
This was caught by the Frame_order.test_opendx_map system test.  The scaling matrix was not being
specified by the dx.map user function backend and this was causing the method to fail.  Instead of
passing the non-existent scaling matrix into the target function, the argument is simply ignored.
The scaling matrix has no effect on the minimise.calculate user function so it is not necessary.