For relaxation dispersion, modified that the Grid search and linear constraints for parameter "k_AB" is between 0-100.
The parameter is only used in the TSMFK01 model.

The "k_AB" parameter is only for very slow forward exchange rate.
The extected values should according to the reference paper:

 Tollinger, M., Skrynnikov, N. R., Mulder, F. A. A., Forman-Kay, J. D., and Kay, L. E. (2001). Slow dynamics in folded and unfolded states of an sh3 domain. J. Am. Chem. Soc., 123(46), 11341-11352. (10.1021/ja011300z)

The paper conserns values of k_AB in the region 0.1 to 5.0.

If the exchange rate is any higher value of this, then another model should be used for the analysis.