Inserted a "RelaxImplementError" when trying to call bmrb_write from a relaxation dispersion analysis.
To implement the fucntion, it would require a re-write of the relax_data bmrb_write(star) function, and proper handling of cdp.ri_ids.
It was also not readily possible to find examples of submitted CPMG data in the BMRB database.
This makes it hard to develop, and even ensure that BMRB would accept the format.

task #7858 (https://gna.org/task/?7858): Make it possible to submit CPMG experiments for BMRB