Improvements for PDB creation in the relax library for out of bounds structural coordinates.
The lib.structure.pdb_write module atom() and hetatm() functions will now more gracefully handle
atomic coordinates which are outside of the PDB limits of [-999.999, 9999.999].  When such
coordinates are encountered, instead of producing a too long PDB line which does not pass the
validation step, the functions will set the coordinates to the boundary value.  This will at least
allow a valid PDB file to be created, despite the warping of the coordinates.