Intermediate optimisation results are now stored by the frame order auto-analysis.
The results from each minimise.grid_search and minimise.execute user function call are now stored in
specially named directories located in the 'intermediate_results' directory, which itself is located
in the auto-analysis results_dir directory.  This allows intermediate results to be more easily
analysed later on, which can be useful for optimising the optimisation steps.  These directories can
also be used for the pre_run_dir auto-analysis argument for subsequent refinements from earlier
steps in the optimisation.  The results stored include everything from the results_output() method
and the count_sobol_points() and summarise() functions.

To allow this to work, the auto-analysis functions count_sobol_points() and summarise() required
modification.  Results files are now always loaded into a temporary data pipe, rather switching to
the corresponding pipe, and the temporary data pipe is deleted after the data has been extracted.
The original data pipe name is also stored and a switch back to that pipe occurs at the end of each
function.