Added instructions and a build script for creating a useful version of latex2html.
This version is essential for building the HTML version of the manual at
http://www.nmr-relax.com/manual/.  The build script downloads the Debian latex2html-2008 sources as
well as all Debian patches for latex2html.  It then applies a number of patches for fixing and
improving the relax documentation.  The program is then compiled and can be installed as the root
user into /usr/local/.