Updated the Palmer.test_palmer_omp system test for the 64-bit Linux Modelfree 4.20 GCC binary file.
This fixes the second part and last part of the bug #22772 (https://gna.org/bugs/?22772), the
modelfree4 binary issue and the User_functions GUI tests with wxPython 2.9 failures of the test
suite.

The problem is that the 64-bit GNU/Linux GCC compiled binary of Modelfree 4.20 produces different
results as previous versions.  These are now caught by the system test and correctly checked.