Fix for the find next bug in the relax controller window.
This is bug #22815 (https://gna.org/bugs/?22815), the failure of find next using F3 (or Ctrl-G on
Mac OS X) in the relax controller window if search text has already been set.

The fix was simple, as the required flags are in the self.find_data class object (an instance of
wx.FindReplaceData).