The splitting of the rigid model grid search into rotation and translation parts is now optional.
In the frame order auto-analysis, the rigid_grid_split argument has been introduced.  The
alternating algorithm of performing a grid search over the rotational space followed by translation
is now optional and turned off by default.  The reason is because the global minimum is sometimes
missed with this shortcut algorithm.