Modified the N_state_model.test_populations system test to test the grid search code paths.
This performs a grid search of one increment after minimisation, then switches to the 'fixed'
N-state model and performs a second grid search of one increment.  This now tests currently untested
code paths in the grid_search() API method behind the minimise.grid_search user function.  The test
demonstrates a bug in the N-state model which was not uncovered in the test suite.