Final fix for bug #22849 (https://gna.org/bugs/?22849).
This is the failure of the N-state model analysis when optimising only alignment tensors using RDCs
and/or PCSs.  The alignment tensor is no longer initialised to zero values.  This is to allow the
skip_preset argument for the minimise.grid_search user function to be operational for the N-state
model, a feature introduced with the zooming grid search.  The solution was to check for the
uninitialised tensor in the minimise_setup_fixed_tensors() method of the
specific_analyses.n_state_model.optimisation module.