Modified the atomic position handling in pipe_control.structure.main.load_spins_multi_mol().
The multiple molecule merging functionality of the structure.load_spins user function now handles
missing atomic positions differently.  The aim is that the length of the spin container position
variable is fixed for all spins to the number of structures, as the N-state model analysis assumes
this equal length for all spins.  When data is missing, the atomic position for that structure is
now set to None.  This will require other modifications in relax to support this new design.