The structure.load_spins user function now stores the number of states in cdp.N.
This is to help the specific analyses which handle ensembles of structures.  With the introduction
of the from_mols argument to the structure.load_spins user function, the number of states is now not
equal to the number of structural models, as the states can now come from different structures of
the same model.  Therefore the user function will now explicitly set cdp.N to the number of states
depending on how the spins were loaded.