Fix for the structure.load_spins user function when the from_mols argument is used.
The load_spins_multi_mol() function of the pipe_control.structure.main module was incorrectly
handling the atomic position returned by the internal structural object atom_loop() method.  This
position is a list of lists when multiple models are present.  But when only a single model is
present, it returns a simple list.