The mol_name argument is now exposed in the structure.add_atom user function.
This has been added as the first argument of the user function to allow new molecules to be created
or to allow the atom to be placed into a specific molecule container.  The functionality was already
implemented in the backend, so it has been exposed by simply adding a new argument definition to the
user function.