Redesigned the optimisation steps for the frame order auto-analysis.
The code has been significantly simplified as the optimisation for the PCS subset and full data set
was the same.  The code duplication has been eliminated by combining it into the new optimisation()
method.  The check for the PCS subset has also been expanded so that it is skipped if the subset
data pipe is not supplied, even if an optimisation object for the subset has been (this should
prevent strange errors when the auto-analysis is incorrectly used).

A side effect of this code merger is that the zooming grid search has now been activated for the
full PCS data set.  This is of great benefit when a PCS subset is not being used.