The pivot point can now be excluded from the grid search in the frame order auto-analysis.
If the initial pivot point is known to be reasonable, then it may be possible to skip it in the grid
search for the rotor frame order model.  This can lead to a speed up of the analysis and can help
with stability.  The pivot_search argument has been added to the auto-analysis
Optimisation.add_grid() method to enable this.  The get_grid_pivot_search() method has also been
added to allow the auto-analysis to query this and turn it off if desired.