Updated frame_order_simulate.py to be much faster in simulating the frame order matrix elements.
The script also matches the Grace file output of the frame_order_solution.py script.  The inside()
method has been renamed for the pseudo-ellipse and the infrastructure for adding support for the
other frame order models has been added.  By shifting calculations outside of the loops, the script
is now many orders of magnitude faster.