Added the isotropic cone model to the frame order matrix element simulation script.
The generated in-frame and out-of-frame Grace graphs for the torsion angle cone_sigma_max,
containing the matrix values for 1,000,000 simulation values have been added to the repository.  The
inside_iso_cone() method has been created to check for the theta_x and theta_z angle violations from
the rotation_hypersphere() method.