Fix for bug #22937 (https://gna.org/bugs/?22937).
This is the failure of the Relax_disp.test_estimate_r2eff_err_auto system test on Python 2.5.  The
test_suite/shared_data/dispersion/Kjaergaard_et_al_2013/1_setup_r1rho_GUI.py simply required a
newline character at the end of the file so that it can be executed in Python 2.5.