Fix for bug #22938 (https://gna.org/bugs/?22938).
This is the failure of the test suite in the relax GUI.  The problem was that the
status.skip_blacklisted_tests variable did not exist - it was only initialised if relax is started
in test suite mode.  Now the value is always set from within the status module and defaults to True.