Bug fix for the structure.load_spins user function new from_mols argument.
This was incorrectly using the pipe_control.pipes.pipe_names() function to obtain its default values
in the GUI (although this is not currently uesd).  The result was a non-fatal error message on Mac
OS X systems of "Python[1065:1d03] *** __NSAutoreleaseNoPool(): Object 0x3a3944c of class NSCFString
autoreleased with no pool in place - just leaking".