Size fix for the rotor representation from the frame_order.pdb_model user function.
The size problem was detected via the Frame_order.test_pdb_model_rotor system test.  The rotors in
the PDB representation were all fixed in size, and ignored the 'size' argument of the
frame_order.pdb_model user function.  The size argument is now passed into the add_rotors() function
of the specific_analyses.frame_order.geometric and passed on to the rotor() function of the
lib.structure.represent.rotor module.