Modified the printouts from the structure.write_pdb user function if models are present.
Instead of printing out 'MODEL', 'ATOM, HETATM, TER' and 'ENDMDL' for each model, the header 'MODEL
records' is printed followed by a single '.' character for each model.  For structures with many models,
this results in a huge speed up of the user function which is strongly limited by how fast the
terminal can display text.