Fix for the Frame_order.test_pdb_model_free_rotor_xz_plane_tilt system test.
This was broken while implementing the Frame_order.test_simulate_rotor_z_axis system test.  Instead
of shifting the 6 atom structure so its CoM is the pivot of the motion when creating the atoms, now
the Frame_order.test_simulate_rotor_z_axis system test sets the average domain translation vector to
the pivot to achieve the same result.  This preserves the z-axis orientation of the rotor models.