Renamed all of the relaxation curve-fitting target functions.
This includes all of the C functions which are model specific, by appending '_exp' to the current
names to now be func_exp, dfunc_exp, d2func_exp, jacobian_exp, and jacobian_chi2_exp.  And all of
the Relax_fit_opt target function class *_wrapper() methods to *_exp().  The target function class
is now only aliasing the *_exp() methods when the model is set to 'exp'.