Creation of a test branch for playing with the idea of a relax_fit C class with numpy support.
This idea has the huge disadvantage in that the compiled C modules will then be specific for one
certain numpy version.  However numpy is constantly changing and therefore targeting one numpy
version is not feasible.  The only solution would be to redesign the distribution archives to bundle
Python, numpy, wxPython, etc., as the Mac OS X DMG file currently does, so that the versions of all
the modules are fixed.